\name{runChains}
\alias{runChains}
\title{
Run MCMC on data
}
\description{
Estimates posterior density for contamination mixture proportions using
Gibbs sampling.
}
\usage{
runChains(data, nChains, nIter, alpha = 0.1, recordAll = FALSE)
}
\arguments{
  \item{data}{
    \code{contamMix} object generated by \code{\link{loadSAM}}
  }
  \item{nChains}{
    Number of independent MCMC chains to run (ideally in parallel).
  }
  \item{nIter}{
    Number of iterations for each chain.
  }
  \item{alpha}{
    Alpha hyperparameter to Dirichlet prior distribution.  This may need
    to be tweaked depending on the data.
  }
  \item{recordAll}{
    Whether to record the samples for all proproportions or only the
    proportion authentic.  
  }
}
\details{
  To run multiple MCMC chains in parallel on a typical machine, set
  \code{options(mc.cores = detectCores())} (for reasons unclear to me,
  the default assumes a single core).
}
\value{
  \code{contamMix} object, which consists of the elements from the input
  \code{data} plus:
  \item{chains}{list of \code{nChains} matrices.  Each matrix has rows
    for each of the \code{nIter} iterations.  The first column contains
    the likelihood, the second contains the proportion authentic, and,
    if recordAll was TRUE, then further columns contain the other
    proportions.
  }
}
\references{
As of May 2013, the best reference is the supplementary information of
Curr Biol. 2013 Apr 8;23(7):553-9.
}
\author{
  Philip Johnson <plfjohnson@emory.edu>
}

\seealso{
  \code{\link{contamMix}}
}
\examples{
\dontrun{
data = loadSAM(samFn="reads.sam", malnFn="aln.fa")
res = runChains(data=data, nChains=3, nIter=10000, alpha=0.1)
print(res)
}
}
\keyword{ optimize }
